+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ```
     * Changes on checkout page logic - when client change the billing address just reload the SDK with the new country, but do not update the Order. The Order will be updated later.
     * Allow plugin Title settings to be set per site view.
+    * Fixed some visual bugs on the checkout page.
 ```
 
 # 2.0.1